Support multidimensional operator[]
This is exactly the same as operator() for views, so not that useful in
ra-ra. Keeping up I suppose.

This support requires -std=c++2b. Unfortunately -std=c++2b breaks operator<<
Big<char, n> on gcc 11 and I don't want to investigate that since gcc 12 works
fine.

The tests for multidimensional [] are feature guarded so this version still
passes all the tests in gcc 11 as long as you compile with -std=c++20 and not
-std=c++2b.

* config/ra.py: Test with -std=c++2b.
* ra/big.hh (operator[]): Forward to ().
* ra/small.hh (operator[]): Forward to ().
* test/big-1.cc: Check multidimensional operator[].
* test/small-1.cc: Check multidimensional operator[].
* test/wrank.cc: Remove some unnecessary printing.