Merge ra::vector in ra::ptr
Mostly to free the ra::vector name, but I think it might be clearer this way.

* ra/atom.hh (ptr): As stated.

Elsewhere replace ra::vector by ra::ptr. Fix some uses of ra::ptr(builtin array)
which now require a cast. These are all checks. In actual user code there should
be no need to erase the size, but if there is, better to be explicit about it.