Patch lookup issues with std::max
This isn't an overload set but a template, so we can't let std::max be inside
ra:: where it matches against our own objects. This only became an issue when
*CellBig started returning View const &, which happens to match std::max
signature.

Didn't think of this before, but std::max isn't like e.g. std::exp or std::cos
in that it's not a variant that applies to certain std:: or pod types, but a
piece of logic meaning a>b?a:b for anything at all. That's not want for arrays.

* ra/ra.hh (DEF_NAME_OP etc): Redo. Use forwarding for std::max, std::min. The
  others as before, because they are overload sets not templates, so they don't
  match against our objects.
* ra/test/reduction-1.cc: Most cases work, but note a new one that doesn't.