Separate Small from SmallView
SmallBase remains for some common types and interfaces, but all the logic is now
in SmallView and Small forwards to that by making views on the spot. This way I
don't need CRTP anymore, the const/nonconst macro block is gone (well, it's
contained), and all of SmallView's members are const, like in View. This should
make it easier to simplify further.

Still, in a way we are diverging from View/Container, because Container remains
a View, like Small was a SmallBase before. And I don't see yet how Container and
View could be separated in the same way since creating a View from a Container
is potentially expensive. Hmm.

* ra/small.hh (unbeat): Make sure to forward op.
  (SmallBase, SmallView, Small): As stated. Follow the pattern of
  View/Container.
  (SmallView::operator()): Copy view to output expr. Should be forward, but copy
  will do until we get explicit this.
  (CellSmall::operator*): Return ctype const &. We can do this bc all SmallView
  methods are const, but also because it matches CellBig's signature.
* test/ra-5.cc: These tests now depend on forwarding in unbeat() and
  SmallView::operator() that were fixed above.