Recover optimization of rank-1 container
This used to break test/io.cc, but now that operator<< has been rewritten, it
doesn't anymore.

This optimization lets the compiler know that the step on this object is 1. It's
equivalent to checking the steps in ply_xxx and doing a special cases for 1 steps,
but that's more code.

* ra/expr.hh (WithLen): Guard against unhandled cases.
* ra/big.hh (Container::iter): Return Ptr rather than CellBig when Ptr is enough.