Alternative prefix matching logic
This always gives a meaningful value for Match::len, even if there is
mismatch. The idea is to check only the root of an expression tree, for instance
at ply(), so that each subexpr's len is only called once, unlike the previous
logic where each subexpr is checked on creation, which results in subexpr's len
being called multiple times.

In progress, the check is still at expr creation.