Fix SmallArray with size 0
The original breakage comes from SmallBase<T, Dimv> a couple commits ago. When I
removed the Dimv dependence, it made the base class of every SmallArray<T> the
same no matter the T, which resulted in a padding bug. But removing the base
class entirely makes sizeof(SmallArray)=0 for size 0, which is an
inconvenience. For example, one can't create a std::vector of such objects.

Options are to 1) use a special object for size 0, or reuse std::array, which
does that; or to 2) add a member or a base class that'll take space only when
the size is 0. But now I know why.

* ra/small.hh (SmallArray): As stated.
* test/sizeof.cc: Add test.

Elsewhere use SmallArray::data() in a few places instead of .cp directly.